using versatileimagefield for images, initialize update quantity form

+++ apps/cart/cart.py
@@ -21,7 +21,7 @@ class Cart:
 
         for item in cart.values():
             item['price'] = Decimal(item['price'])
-            item['total_price'] = item['price'] * item['quantity']
+            item['total_price'] = Decimal(item['price'] * item['quantity'])
             yield item
 
     def __len__(self):
